Implemented a multi stage Docker build to build a docker image and run the app in a container. However, I faced a big issue setting up the integration test to execute in the container.
For now, the tests are skipped as part of the Docker build and can be executed only manually via Maven.
Will try to fix the Integration test as part of the Docker build on a later stage.

Testing done:
run `docker-compose up -build` to run the app in a Docker container
run `mvnw clean install` to verify the app is build via maven too and the tests execute successfully